+++ android-exoplayer/src/main/java/com/brentvatne/exoplayer/ReactExoplayerView.java
@@ -81,6 +81,8 @@ import java.util.ArrayList;
 import java.util.Locale;
 import java.util.UUID;
 import java.util.Map;
+import java.util.Timer;
+import java.util.TimerTask;
 
 @SuppressLint("ViewConstructor")
 class ReactExoplayerView extends FrameLayout implements
@@ -137,6 +139,7 @@ class ReactExoplayerView extends FrameLayout implements
     private int bufferForPlaybackAfterRebufferMs = DefaultLoadControl.DEFAULT_BUFFER_FOR_PLAYBACK_AFTER_REBUFFER_MS;
 
     private Handler mainHandler;
+    private Timer bufferCheckTimer;
 
     // Props from React
     private int backBufferDurationMs = DefaultLoadControl.DEFAULT_BACK_BUFFER_DURATION_MS;
@@ -415,6 +418,40 @@ class ReactExoplayerView extends FrameLayout implements
         }
     }
 
+    private void startBufferCheckTimer() {
+        SimpleExoPlayer player = this.player;
+        VideoEventEmitter eventEmitter = this.eventEmitter;
+        Handler mainHandler = this.mainHandler;
+
+        if (this.bufferCheckTimer != null) {
+            this.stopBufferCheckTimer();
+        }
+
+        this.bufferCheckTimer = new Timer();
+        TimerTask bufferCheckTimerTask = new TimerTask() {
+            @Override
+            public void run() {
+                if (mainHandler != null) {
+                    mainHandler.post(new Runnable() {
+                        public void run() {
+                            if (player != null) {
+                                double bufferedDuration = (double) (player.getBufferedPercentage() * player.getDuration() / 100);
+                                eventEmitter.bufferProgress(0d, bufferedDuration);
+                            }
+                        }
+                    });
+                }
+            };
+        };
+
+        this.bufferCheckTimer.scheduleAtFixedRate(bufferCheckTimerTask, 500, 1000);
+    }
+
+    private void stopBufferCheckTimer() {
+        this.bufferCheckTimer.cancel();
+        this.bufferCheckTimer = null;
+    }
+
     private void initializePlayer() {
         ReactExoplayerView self = this;
         // This ensures all props have been settled, to avoid async racing conditions.
@@ -506,6 +543,7 @@ class ReactExoplayerView extends FrameLayout implements
                 initializePlayerControl();
                 setControls(controls);
                 applyModifiers();
+                startBufferCheckTimer();
             }
         }, 1);
     }
@@ -597,6 +635,7 @@ class ReactExoplayerView extends FrameLayout implements
 
     private void releasePlayer() {
         if (player != null) {
+            stopBufferCheckTimer();
             updateResumePosition();
             player.release();
             player.removeMetadataOutput(this);